+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v2/caching/agent/KubernetesNamespaceCachingAgent.java
@@ -53,10 +53,8 @@ public class KubernetesNamespaceCachingAgent extends KubernetesV2CachingAgent {
 
   @Override
   protected Map<KubernetesKind, List<KubernetesManifest>> loadPrimaryResourceList() {
-    reloadNamespaces();
-
     // TODO perf: Only load desired namespaces rather than filter all.
-    Set<String> desired = new HashSet<>(this.namespaces);
+    Set<String> desired = new HashSet<>(this.getNamespaces());
     return Collections.singletonMap(
         KubernetesKind.NAMESPACE,
         credentials.list(KubernetesKind.NAMESPACE, "").stream()
